+++ packages/calendar/src/usePeriod.ts
@@ -6,11 +6,25 @@ type usePeriodProps = {
      * Обработчик изменения выделенного периода
      */
     onPeriodChange?: (selectedFrom?: number, selectedTo?: number) => void;
+
+    /**
+     * Начальное значение начала выделенного периода
+     */
+    initialSelectedFrom?: number;
+
+    /**
+     * Начальное значение конца выделенного периода
+     */
+    initialSelectedTo?: number;
 };
 
-export function usePeriod({ onPeriodChange }: usePeriodProps) {
-    const [selectedFrom, setSelectedFrom] = useState<number>();
-    const [selectedTo, setSelectedTo] = useState<number>();
+export function usePeriod({
+    onPeriodChange,
+    initialSelectedFrom,
+    initialSelectedTo,
+}: usePeriodProps) {
+    const [selectedFrom, setSelectedFrom] = useState<number | undefined>(initialSelectedFrom);
+    const [selectedTo, setSelectedTo] = useState<number | undefined>(initialSelectedTo);
 
     const resetPeriod = useCallback(() => {
         setSelectedFrom(undefined);