+++ packages/select/src/components/options-list/Component.tsx
@@ -19,6 +19,7 @@ export const OptionsList = ({
     options = [],
     Optgroup = DefaultOptgroup,
     dataTestId,
+    emptyPlaceholder,
 }: OptionsListProps) => {
     const counter = createCounter();
 
@@ -31,11 +32,19 @@ export const OptionsList = ({
         [Option, counter, size],
     );
 
-    return options.length > 0 ? (
+    if (options.length === 0 && !emptyPlaceholder) {
+        return null;
+    }
+
+    return (
         <div className={cn(styles.optionsList, styles[size], className)} data-test-id={dataTestId}>
             {options.map(option =>
                 isGroup(option) ? renderGroup(option) : Option({ option, index: counter() }),
             )}
+
+            {emptyPlaceholder && options.length === 0 && (
+                <div className={styles.emptyPlaceholder}>{emptyPlaceholder}</div>
+            )}
         </div>
-    ) : null;
+    );
 };